Move inquiries route handling to separate file
