+++ src/apps/loyalty/__mocks__/isomorphic-relay.ts
@@ -1,0 +1,3 @@
+module.exports = {
+  container: container => container
+}
