+++ src/apps/loyalty/server/index.tsx
@@ -1,22 +1,13 @@
 import * as artsyPassport from "artsy-passport"
 import * as Backbone from "backbone"
 import * as express from "express"
-import { default as IsomorphicRelay } from "isomorphic-relay"
 import * as path from "path"
 import * as React from "react"
 import * as Relay from "react-relay"
 
-import { renderToString } from "react-dom/server"
-import * as styleSheet from "styled-components/lib/models/StyleSheet"
-import renderPage from "./template"
-
-import CurrentUserRoute from "../../../relay/queries/current_user"
-
-import Inquiries from "../containers/inquiries"
-import { markCollectorAsLoyaltyApplicant } from "./gravity"
 import RelayMiddleware from "./middlewares/relay"
 import UserMiddleware from "./middlewares/user"
-import { Home, Login, ThankYou } from "./route_handlers"
+import { Home, Inquiries, Login, ThankYou } from "./route_handlers"
 
 const app = express.Router()
 const { API_URL } = process.env
@@ -38,42 +29,6 @@ app.use(UserMiddleware)
 app.get("/", Home)
 app.get(loginPagePath, Login)
 app.get("/thank-you", ThankYou)
-
-app.get("/inquiries", (req, res) => {
-  if (!req.user) {
-    return res.redirect(req.baseUrl + loginPagePath)
-  }
-
-  const info = req.user.get("profile")
-  if (info.loyalty_applicant_at) {
-    return res.redirect(req.baseUrl + "/thank-you")
-  }
-
-  if (info.confirmed_buyer_at) {
-    markCollectorAsLoyaltyApplicant(req.user.get("accessToken"))
-      .then(profile => {
-        return res.redirect(req.baseUrl + "/thank-you")
-      })
-      .catch(err => console.error(err))
-  }
-
-  IsomorphicRelay.prepareData({
-    Container: Inquiries,
-    queryConfig: new CurrentUserRoute(),
-  }, res.locals.networkLayer).then(
-    ({data, props}) => {
-      const html = renderToString(<IsomorphicRelay.Renderer {...props} />)
-      const styles = styleSheet.rules().map(rule => rule.cssText).join("\n")
-      res.locals.sharify.data.USER_DATA = req.user.toJSON()
-      res.locals.sharify.data.DATA = data
-      res.send(renderPage({
-        styles,
-        html,
-        entrypoint: req.baseUrl + "/bundles/inquiries.js",
-        sharify: res.locals.sharify.script(),
-        baseURL: req.baseUrl,
-      }))
-    })
-})
+app.get("/inquiries", Inquiries)
 
 export default app
