+++ src/apps/loyalty/server/route_handlers.tsx
@@ -1,5 +1,6 @@
 import * as artsyPassport from "artsy-passport"
 import { NextFunction, Request, Response } from "express"
+import { default as IsomorphicRelay } from "isomorphic-relay"
 import * as React from "react"
 
 import { renderToString } from "react-dom/server"
@@ -8,8 +9,13 @@ import LoginContainer from "../containers/login"
 import { CollectorProfileResponse } from "./gravity"
 import renderPage from "./template"
 
+import CurrentUserRoute from "../../../relay/queries/current_user"
+
 import ThreewThankYou from "../containers/3w_thank_you"
 import AcbThankYou from "../containers/acb_thank_you"
+import InquiriesContainer from "../containers/inquiries"
+
+import { markCollectorAsLoyaltyApplicant } from "./gravity"
 
 export function Home(req: Request, res: Response, next: NextFunction) {
   return res.redirect(req.baseUrl + "/inquiries")
@@ -39,6 +45,44 @@ export function ThankYou(req: Request, res: Response, next: NextFunction) {
 
   const styles = styleSheet.rules().map(rule => rule.cssText).join("\n")
   return res.send(renderPage({ styles, html, entrypoint: "", baseURL: req.baseUrl }))
+}
+
+export function Inquiries(req: Request, res: Response, next: NextFunction) {
+  const { loginPagePath } = artsyPassport.options
+  if (!req.user) {
+    return res.redirect(req.baseUrl + loginPagePath)
+  }
+
+  const info = req.user.get("profile")
+  if (info.loyalty_applicant_at) {
+    return res.redirect(req.baseUrl + "/thank-you")
+  }
+
+  if (info.confirmed_buyer_at) {
+    markCollectorAsLoyaltyApplicant(req.user.get("accessToken"))
+      .then(profile => {
+        return res.redirect(req.baseUrl + "/thank-you")
+      })
+      .catch(err => console.error(err))
+  }
+
+  IsomorphicRelay.prepareData({
+    Container: InquiriesContainer,
+    queryConfig: new CurrentUserRoute(),
+  }, res.locals.networkLayer).then(
+    ({data, props}) => {
+      const html = renderToString(<IsomorphicRelay.Renderer {...props} />)
+      const styles = styleSheet.rules().map(rule => rule.cssText).join("\n")
+      res.locals.sharify.data.USER_DATA = req.user.toJSON()
+      res.locals.sharify.data.DATA = data
+      res.send(renderPage({
+        styles,
+        html,
+        entrypoint: req.baseUrl + "/bundles/inquiries.js",
+        sharify: res.locals.sharify.script(),
+        baseURL: req.baseUrl,
+      }))
+    })
 }
 
 export function Login(req: Request, res: Response, next: NextFunction) {
